Refaktoroi: Siirrä pitkän testin helpperit omaan tiedostoon;
Testitiedosto on paisunut niin isoksi, että koodieditori alkaa hyytyä sen
parsimiseen. Helpota tuskaa siirtämällä osa kompleksisista metodeista
kantaluokkaan.
